hstd: specify edge checkpoints in constraints
