+++ examples/imgui_gui/block_graph.cpp
@@ -10,92 +10,6 @@ struct DocLayout {
     UnorderedMap<DocNode, int> rectMap;
 };
 
-template <typename T>
-struct JsonSerde<Slice<T>> {
-    static T from_json(json const& j) {
-        return slice<T>(
-            JsonSerde<T>::from_json(j["first"]),
-            JsonSerde<T>::from_json(j["last"]));
-    }
-
-    static json to_json(Slice<T> const& value) {
-        return json::object({
-            {"first", JsonSerde<T>::to_json(value.first)},
-            {"last", JsonSerde<T>::to_json(value.last)},
-        });
-    }
-};
-
-template <typename A, typename B>
-struct JsonSerde<Pair<A, B>> {
-    static Pair<A, B> from_json(json const& j) {
-        return std::make_pair<A, B>(
-            JsonSerde<A>::from_json(j["first"]),
-            JsonSerde<B>::from_json(j["second"]));
-    }
-
-    static json to_json(Pair<A, B> const& value) {
-        return json::object({
-            {"first", JsonSerde<A>::to_json(value.first)},
-            {"second", JsonSerde<B>::to_json(value.second)},
-        });
-    }
-};
-
-template <typename K, typename V>
-struct JsonSerde<UnorderedMap<K, V>> {
-    static json to_json(UnorderedMap<K, V> const& it) {
-        auto result = json::array();
-        for (auto const& [key, val] : it) {
-            result.push_back(json::object({
-                {"key", JsonSerde<K>::to_json(key)},
-                {"value", JsonSerde<V>::to_json(val)},
-            }));
-        }
-
-        return result;
-    }
-    static UnorderedMap<K, V> from_json(json const& j) {
-        UnorderedMap<K, V> result;
-        auto               tmp = result.transient();
-        for (auto const& i : j) {
-            result.insert(
-                JsonSerde<K>::from_json(i["key"]),
-                JsonSerde<V>::from_json(i["value"]));
-        }
-        return tmp.persistent();
-    }
-};
-
-template <IsVariant V>
-struct JsonSerde<V> {
-    static json to_json(V const& it) {
-        auto result     = json::object();
-        result["index"] = JsonSerde<int>::to_json(it.index());
-        std::visit(
-            [&]<typename T>(T const& value) {
-                result["value"] = JsonSerde<T>::to_json(value);
-            },
-            it);
-
-        return result;
-    }
-    static V from_json(json const& j) {
-        V result = variant_from_index<V>(j["index"].get<int>());
-        std::visit(
-            [&]<typename T>(T& value) {
-                value = JsonSerde<T>::from_json(j["value"]);
-            },
-            result);
-    }
-};
-
-template <>
-struct JsonSerde<double> {
-    static json   to_json(double const& it) { return json(it); }
-    static double from_json(json const& j) { return j.get<double>(); }
-};
-
 
 GraphNodeConstraint::Align::Spec spec(int rect) {
     return GraphNodeConstraint::Align::Spec{.node = rect};
