+++ src/hstd/wrappers/adaptagrams_wrap/adaptagrams_ir.hpp
@@ -69,7 +69,7 @@ struct [[refl]] GraphRect {
 
     bool contains(GraphPoint const& p) const {
         return (left <= p.x && p.x <= (left + width))
-            && (top <= p.y && p.y <= (top + height));
+        && (top <= p.y && p.y <= (top + height));
     }
 
     void extend(GraphPoint const& point);
@@ -84,8 +84,8 @@ struct [[refl]] GraphNodeConstraint {
         DESC_FIELDS(Empty, ());
     };
 
-    /// \brief Listed nodes must be positioned on the same X/Y dimension in
-    /// the layout
+           /// \brief Listed nodes must be positioned on the same X/Y dimension in
+           /// the layout
     struct [[refl]] Align {
         struct [[refl]] Spec {
             [[refl]] int         node; ///< Rectangle index
@@ -98,7 +98,7 @@ struct [[refl]] GraphNodeConstraint {
         [[refl]] GraphDimension dimension; ///< Which axist to align on
         DESC_FIELDS(Align, (nodes, dimension));
 
-        /// \brief Map to cola layout constraint object
+               /// \brief Map to cola layout constraint object
         Res                  toCola() const;
         [[refl]] std::string toColaString() const {
             return toCola()->toString();
@@ -309,6 +309,16 @@ struct [[refl]] GraphEdgeConstraint {
 
     [[refl]] Port sourcePort;
     [[refl]] Port targetPort;
+    /// \brief Offset the port position on the side of the shape. Value in
+    /// range [0, 1] used as a relative position along the edge of the
+    /// shape. For vertical edges (west/east) the port movement would be
+    /// vertical, and for horizontal (north/south) it would be horizontal.
+    /// If the value is not provided the nodes are placed in the center
+    /// with the ratio computed automatically.
+    [[refl]] Opt<double> sourceOffset     = std::nullopt;
+    [[refl]] Opt<double> targetOffset     = std::nullopt;
+    [[refl]] Opt<double> sourceCheckpoint = 10.0;
+    [[refl]] Opt<double> targetCheckpoint = 10.0;
 
     DESC_FIELDS(GraphEdgeConstraint, (sourcePort, targetPort));
 };
@@ -323,9 +333,9 @@ struct [[refl]] GraphLayoutIR {
     struct [[refl]] Subgraph {
         Str graphName; ///< Graphviz graph name
 
-        /// \brief Which nodes go directly in this cluster level (subgraphs
-        /// will have their own sub-nodes, there should be no overlap or
-        /// duplication)
+               /// \brief Which nodes go directly in this cluster level (subgraphs
+               /// will have their own sub-nodes, there should be no overlap or
+               /// duplication)
         Vec<int> nodes;
         /// \brief List of nested subgraphs
         Vec<Subgraph> subgraphs;
@@ -340,9 +350,9 @@ struct [[refl]] GraphLayoutIR {
         bool isEmpty() const;
     };
 
-    /// \brief Nodes for the graph. Node is identified by the index in the
-    /// array of sizes. In the result value each original qsize is mapped
-    /// to the rectangle.
+           /// \brief Nodes for the graph. Node is identified by the index in the
+           /// array of sizes. In the result value each original qsize is mapped
+           /// to the rectangle.
     [[refl]] Vec<GraphSize> rectangles;
     /// \brief List of source-target pairs. Edge source/target IDs refer to
     /// the size rectangles.
@@ -369,22 +379,21 @@ struct [[refl]] GraphLayoutIR {
         (rectangles,
          edges,
          nodeConstraints,
-         edgeConstraints,
          subgraphs,
          edgeLabels,
          width,
          height,
          graphName));
 
-    /// \brief Which DPI to use when converting to and from graphviz sizes.
-    /// Backend-specific, 72 is the default used by graphviz.
+           /// \brief Which DPI to use when converting to and from graphviz sizes.
+           /// Backend-specific, 72 is the default used by graphviz.
     [[refl]] int graphviz_size_scaling = 72;
 
-    /// \brief validate the edge/rectangle structure for debugging. Throws
-    /// assert failure if the structure is incorrect.
+           /// \brief validate the edge/rectangle structure for debugging. Throws
+           /// assert failure if the structure is incorrect.
     [[refl]] void validate();
 
-    /// \brief Graph edge layout
+           /// \brief Graph edge layout
     struct [[refl]] Edge {
         /// \brief Sequence of painter paths going from source to target
         /// node. If the node has a label rectangle specified, the paths
@@ -395,7 +404,7 @@ struct [[refl]] GraphLayoutIR {
         DESC_FIELDS(Edge, (paths, labelRect));
     };
 
-    /// \brief Full layout result from the conversion
+           /// \brief Full layout result from the conversion
     struct [[refl]] Result {
         /// \brief Recursive subgraph layout. Nodes and edges inside of the
         /// subgraph are stored in a flat layout result.
@@ -406,7 +415,7 @@ struct [[refl]] GraphLayoutIR {
 
             DESC_FIELDS(Subgraph, (bbox, subgraphs));
 
-            /// \brief Get reference to subgraph specified at path
+                   /// \brief Get reference to subgraph specified at path
             Subgraph const& getSubgraph(Span<int> path) const {
                 switch (path.size()) {
                     case 0: return *this;
@@ -418,9 +427,9 @@ struct [[refl]] GraphLayoutIR {
             }
         };
 
-        /// \brief Fixed node layout rectangles with absolute coordinates.
-        /// Subgraph nodes are also included. Edge label nodes are not
-        /// included.
+               /// \brief Fixed node layout rectangles with absolute coordinates.
+               /// Subgraph nodes are also included. Edge label nodes are not
+               /// included.
         [[refl]] Vec<GraphRect> fixed;
         /// \brief Mapping from the source-target edge pair to the edge
         /// layout spec
@@ -436,7 +445,7 @@ struct [[refl]] GraphLayoutIR {
             Result,
             (fixed, lines, bbox, subgraphs, subgraphPaths));
 
-        /// \brief Get subgraph at path
+               /// \brief Get subgraph at path
         Subgraph const& getSubgraph(Vec<int> const& path) {
             switch (path.size()) {
                 case 0:
@@ -451,7 +460,7 @@ struct [[refl]] GraphLayoutIR {
         }
     };
 
-    /// \brief Backend-specific layout results for graphviz graph
+           /// \brief Backend-specific layout results for graphviz graph
     struct GraphvizResult {
         Graphviz::Graph graph;
         Graphviz        gvc;
@@ -488,7 +497,7 @@ struct [[refl]] GraphLayoutIR {
     HolaResult      doHolaLayout();
     [[refl]] Result doHolaConvert() { return doHolaLayout().convert(); }
 
-    /// \brief Backend-specific layout results for cola graph layout
+           /// \brief Backend-specific layout results for cola graph layout
     struct ColaResult {
         GraphLayoutIR*        baseIr;
         Vec<vpsc::Rectangle>  baseRectangles;
@@ -510,7 +519,7 @@ struct [[refl]] GraphLayoutIR {
 
         Result convert();
 
-        /// \brief write graph layout result into the SVG svg file
+               /// \brief write graph layout result into the SVG svg file
         void writeSvg(CR<Str> path);
 
         Vec<SPtr<cola::CompoundConstraint>> setupConstraints(
